[sailfishos][embedlite] Adapt code from EmbedLiteViewBaseChild/Parent to EmbedLiteViewChild/Parent. JB#54983 OMP#JOLLA-261
The EmbedLiteViewBaseChild/Parent classes previously inherited by
EmbedLiteViewChild/Parent have now been integrated into
EmbedLiteViewChild/Parent. This shifts various changes made to the
former into the latter, specifically to support:

1. Setting desktop mode.
2. Setting a customer HTTP user agent.
3. Handling the virtual keyboard margin.
